modificacion del tipo id para los usuarios en los manejos de seguridad

+++ src/main/java/com/undec/corralon/config/AuditingConfig.java
@@ -18,15 +18,15 @@ import java.util.Optional;
 public class AuditingConfig {
 
     @Bean
-    public AuditorAware<Integer> auditorProvider() {
+    public AuditorAware<Long> auditorProvider() {
         return new SpringSecurityAuditAwareImpl();
     }
 }
 
-class SpringSecurityAuditAwareImpl implements AuditorAware<Integer> {
+class SpringSecurityAuditAwareImpl implements AuditorAware<Long> {
 
     @Override
-    public Optional<Integer> getCurrentAuditor() {
+    public Optional<Long> getCurrentAuditor() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
         if (authentication == null ||
